Rename h4 to h3 and reclassify parallel search as baseline
Renumber h4 (chain congestion + exchange reliability) to h3 so the
three A*-modifying heuristics are h1, h2, h3. Reclassify the old h3
(multi-start parallel search) as a "Parallelized Baseline" that is no
longer numbered as a heuristic.

Changes span 23 files:
- LaTeX: updated all 6 .tex files (abstract, problem formulation,
  novel heuristics, experiments, conclusion, appendix) and recompiled PDF
- Python modules: git mv h3_parallel -> parallel_baseline,
  h4_chaincongestion_exchange_risk -> h3_chaincongestion_exchange_risk
- Experiment files: renamed run_h4_unit_tests -> run_h3_unit_tests,
  run_h3_unit_tests -> run_parallel_baseline_unit_tests
- Updated all imports, string identifiers, dispatch logic, and comments
  across scripts/, experiments/, testing/, and analysis/
- Added HEURISTIC_LABEL_MAP in generate_paper_figures.py so historical
  result files render with the new names

Made-with: Cursor
