+++ docs/latex/StablecoinArbitrage_GSS2026/sec/0_abstract.tex
@@ -1,3 +1,3 @@
 \begin{abstract}
-Cross-exchange cryptocurrency arbitrage enables low-risk profit from price discrepancies across exchanges, yet existing approaches employ negative cycle detection that targets opportunity identification rather than execution feasibility. We introduce an execution-aware pathfinding framework using A* search with domain-specific guidance heuristics, applied to stablecoins, a unique asset class exceeding \$300 billion in market capitalization that bridges cryptocurrency and fiat currency, offering a novel dataset for arbitrage research. The problem is modelled as a weighted directed graph where nodes represent (exchange, stablecoin) pairs across 12 centralized exchanges and edges encode real-world costs including fees, slippage, gas, transfer delays, and exchange reliability. Three guidance heuristics and a multi-start strategy are evaluated over 7{,}200 search instances. Our slippage-aware heuristic $h_2$ reduces node expansions by 29\% relative to Dijkstra while matching its profit, demonstrating that domain-specific heuristics can meaningfully improve execution feasibility in real-time arbitrage planning. Code: \url{https://github.com/kevinl03/Stablecoin-CrossExchange-Arbitrage}.
+Cross-exchange cryptocurrency arbitrage enables low-risk profit from price discrepancies across exchanges, yet existing approaches employ negative cycle detection that targets opportunity identification rather than execution feasibility. We introduce an execution-aware pathfinding framework using A* search with domain-specific guidance heuristics, applied to stablecoins, a unique asset class exceeding \$300 billion in market capitalization that bridges cryptocurrency and fiat currency, offering a novel dataset for arbitrage research. The problem is modelled as a weighted directed graph where nodes represent (exchange, stablecoin) pairs across 12 centralized exchanges and edges encode real-world costs including fees, slippage, gas, transfer delays, and exchange reliability. Three guidance heuristics and a parallelized multi-start baseline are evaluated over 7{,}200 search instances. Our slippage-aware heuristic $h_2$ reduces node expansions by 29\% relative to Dijkstra while matching its profit, demonstrating that domain-specific heuristics can meaningfully improve execution feasibility in real-time arbitrage planning. Code: \url{https://github.com/kevinl03/Stablecoin-CrossExchange-Arbitrage}.
 \end{abstract}
