+++ docs/latex/StablecoinArbitrage_GSS2026/sec/3_problem_formulation.tex
@@ -19,4 +19,4 @@ The goal set is $\mathcal{G} = \bigl\{\, v_k \in V \;\big|\; C_k > C_0 \,\bigr\}
 
 We apply A* search with cost-to-reach $g(n) = \sum_{e\in \text{path}} -\log(r(e))$ and evaluation function $f(n) = g(n) + h(n)$, where $h(n)$ estimates the execution risk of the remaining path given current blockchain state, including factors such as liquidity depth, order-book slippage, chain congestion, and exchange reliability. The algorithm terminates when $C_n > C_0$.
 
-For $h_4$, we employ \textbf{Weighted A*}\textsuperscript{\cite{pohl1970weighted}} with $f(n) = g(n) + w(n) \cdot h(n)$, where $w(n) = w_{\min} + (w_{\max} - w_{\min}) \cdot \rho(n)$ and $\rho(n) \in [0,1]$ is a chain kickback risk score. When $w(n)=1$ for all nodes, this reduces to standard A*. The algorithm is not complete: market data does not guarantee a profitable path from every starting node. Search terminates when a profitable path is found, the frontier is exhausted, or depth/time limits are reached.
+For $h_3$, we employ \textbf{Weighted A*}\textsuperscript{\cite{pohl1970weighted}} with $f(n) = g(n) + w(n) \cdot h(n)$, where $w(n) = w_{\min} + (w_{\max} - w_{\min}) \cdot \rho(n)$ and $\rho(n) \in [0,1]$ is a chain kickback risk score. When $w(n)=1$ for all nodes, this reduces to standard A*. The algorithm is not complete: market data does not guarantee a profitable path from every starting node. Search terminates when a profitable path is found, the frontier is exhausted, or depth/time limits are reached.
