+++ docs/latex/StablecoinArbitrage_GSS2026/sec/6_Conclusion.tex
@@ -1,1 +1,1 @@
-We presented an execution-aware A* framework for cross-exchange stablecoin arbitrage across 12 exchanges (41~nodes, 864~edges). $h_2$ (slippage) reduces expansions by 29\% while matching Dijkstra's profit; $h_1$ and $h_4$ overcompensate risk, expanding more nodes for less profit. Future work includes asynchronous order-book pre-fetching and extension to DEX/AMM markets.
+We presented an execution-aware A* framework for cross-exchange stablecoin arbitrage across 12 exchanges (41~nodes, 864~edges). $h_2$ (slippage) reduces expansions by 29\% while matching Dijkstra's profit; $h_1$ and $h_3$ overcompensate risk, expanding more nodes for less profit. Future work includes asynchronous order-book pre-fetching and extension to DEX/AMM markets.
